(Test): Update tests to reflect new report and dahsboard layout with the bars and all [tsv22]
