+++ packages/frontend/src/app/components/title.tsx
@@ -7,50 +7,6 @@ import { pdf } from '@react-pdf/renderer';
 import ReportRenderer from './ReportRenderer';
 
 
-// const mockReport = {
-//   "info" : {
-//     calls_to_model: 10,
-//     date: "2021-10-10",
-//   },
-
-//   "properties": [
-//     {
-//     "property": "property1",
-//     "computed_metrics": [
-//       {
-//         "metric": "metric1",
-//         "value": "0.5",
-//         "ideal_value": "0.8",
-//         "range": ["0", "1"],
-//       },
-//       {
-//         "metric": "metric2",
-//         "value": "0.7",
-//         "ideal_value": "0.3",
-//         "range": ["-1", "1"],
-//       }],
-//     "legislation_extracts": [
-//       {
-//         "article_number": 1,
-//         "article_title": "title1",
-//         "description": "description1",
-//         "suitable_recitals": ["recital1", "recital2"],
-//       },
-//       {
-//         "article_number": 2,
-//         "article_title": "title2",
-//         "description": "description2",
-//         "suitable_recitals": ["recital3", "recital4"],
-//       }],
-//     "llm_insights": [
-//       "insight1",
-//       "insight2",
-//     ],
-//     },
-//   ]
-// } 
-
-
 const Title = () => (
   <Box style={styles.container}>
     <Box
@@ -83,22 +39,6 @@ const Title = () => (
       >
         Getting Started
       </Button>
-      {/* <Button
-        variant="contained"
-        sx={{
-          ml: '10px',
-          bgcolor: theme.palette.secondary.main,
-          color: '#fff',
-        }}
-        onClick={async () => {
-          const blob = await pdf(<ReportRenderer report={mockReport} />).toBlob();
-          const url = URL.createObjectURL(blob);
-          window.open(url, '_blank');
-        }}
-      >
-        Generate Report
-      </Button> */}
-        
     </Box>
   </Box>
 );
